requirements.txt
aiogram~=2.15
requests~=2.26.0
Pillow~=8.3.2
bs4~=0.0.1
 tk
beautifulsoup4~=4.10.0